pyproject.toml
[project]
name = "rtsp-vision-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.12",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.2.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "pandas>=3.0.1",
    "django-cors-headers>=4.9.0",
    "celery>=5.6.2",
    "django-celery-beat>=2.9.0",
]
